pyproject.toml
[project]
name = "pushqueue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["aerich>=0.8.2", "fastapi>=0.115.12", "httpx>=0.28.1", "tomlkit>=0.13.2", "tortoise-orm>=0.25.0", "uvicorn[standard]>=0.34.2"]

[tool.aerich]
tortoise_orm = "settings.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."
